Task_Django/requirements.txt
Django==5.0.4
djangorestframework==3.15.1
mysqlclient==2.2.4